ReactDOMEventListener: clean up module
Fix
